+++ tests/testthat/test-emmeans.R
@@ -28,3 +28,7 @@ test_that("Nested EMMs work", {
     expect_equal(colnames(emmeans(rgg, ~ source)@grid)[1:2], c("source","group"))
 })
 
+test_that("Column names in linfct are preserved",
+          expect_equal(colnames(emmeans(rg, "source")@linfct),
+                       colnames(rg@linfct))
+)